Make the FileVersion and PutUserInfo test groups required for prod
The FileVersion tests are easy to address for hosts (add a header to the response for several WOPI operations) and reduce the amount of manual testing that needs to be done during validation.

The PutUserInfo tests are new; hosts that support PutUserInfo must be passing these tests to ensure their implementation is correct.